Add files via upload
It is a file from MySQL. This is a project to analyse a simulation from Company Magist. I use the data to an analyse the buisness situation of the company and try to find the development of the company. There is another potential cooperation between this company and another high-tech company Eniac. The main object is to find whether it is an good option to do the cooperation in behalf of Eniac. There is also a small presentation for the analysis like https://docs.google.com/presentation/d/19J2iQijDioA7HaC3ljr6oRCAMNlKM2rMaPtQcarJn54/edit#slide=id.g130521067c1_0_1